+++ Editor/Scripts/Utils/GUIVariableUtils.cs
@@ -8,52 +8,46 @@ namespace Dash.Editor
 {
     public class GUIVariableUtils
     {
-        public static void DrawVariablesInspector(string p_title, DashVariables p_variables, IVariableBindable p_bindable)
+        public static bool DrawVariablesInspector(string p_title, DashVariables p_variables, IVariableBindable p_bindable, float p_maxWidth, ref bool p_minimized)
         {
-            var style = new GUIStyle();
-            style.normal.textColor = new Color(1, 0.7f, 0);
-            style.alignment = TextAnchor.MiddleCenter;
-            style.fontStyle = FontStyle.Bold;
-            style.normal.background = Texture2D.whiteTexture;
-            style.fontSize = 16;
-            GUI.backgroundColor = new Color(0, 0, 0, .5f);
-            GUILayout.Label(p_title, style, GUILayout.Height(28));
-            GUI.backgroundColor = Color.white;
+            if (!GUIEditorUtils.DrawMinimizableSectionTitle(p_title,
+                    ref p_minimized))
+                return false;
+
+            return DrawVariablesInspector("", p_variables, p_bindable, p_maxWidth);
+        }
+        
+        public static bool DrawVariablesInspector(string p_title, DashVariables p_variables, IVariableBindable p_bindable, float p_maxWidth)
+        {
+            if (p_title != "") GUIEditorUtils.DrawSectionTitle(p_title);
 
             int index = 0;
+            bool invalidate = false;
             p_variables.variables?.ForEach(variable =>
             {
-                VariableField(p_variables, variable.Name, p_bindable,
-                    EditorGUIUtility.currentViewWidth - 20);
+                invalidate = invalidate || VariableField(p_variables, variable.Name, p_bindable,
+                    p_maxWidth);
                 GUILayout.Space(4);
                 index++;
             });
 
-            if (GUILayout.Button("Add Variable"))
-            {
-                VariableTypesMenu.Show((type) => OnAddNewVariable(p_variables, type));
-            }
+            return invalidate;
         }
-        
-        static void OnAddNewVariable(DashVariables p_variables, Type p_type)
+
+        public static bool VariableField(DashVariables p_variables, string p_name, IVariableBindable p_bindable, float p_maxWidth)
         {
-            p_variables.AddNewVariable(p_type);
-            // EditorUtility.SetDirty(target);
-            // PrefabUtility.RecordPrefabInstancePropertyModifications(target);
-        }
-        
-        public static void VariableField(DashVariables p_variables, string p_name, IVariableBindable p_bindable, float p_maxWidth)
-        {
+            bool invalidate = false;
             var variable = p_variables.GetVariable(p_name);
             GUILayout.BeginHorizontal();
             string newName = EditorGUILayout.TextField(p_name, GUILayout.Width(120));
             GUILayout.Space(2);
-            if (newName != p_name) 
+            if (newName != p_name)
             {
+                invalidate = true;   
                 p_variables.RenameVariable(p_name, newName);
             }
             
-            variable.ValueField(p_maxWidth-150, p_bindable);
+            invalidate = invalidate || variable.ValueField(p_maxWidth-150, p_bindable);
 
             var oldColor = GUI.color;
             GUI.color = variable.IsBound || variable.IsLookup ? Color.yellow : Color.gray;
@@ -61,7 +55,7 @@ namespace Dash.Editor
             GUILayout.FlexibleSpace();
             GUILayout.BeginVertical(GUILayout.Width(16));
             GUILayout.Space(2);
-            if (GUILayout.Button(IconManager.GetIcon("bind_icon"), GUIStyle.none, GUILayout.Height(16), GUILayout.Width(16)))
+            if (GUILayout.Button(IconManager.GetIcon("settings_icon"), GUIStyle.none, GUILayout.Height(16), GUILayout.Width(16)))
             {
                 var menu = VariableSettingsMenu.Get(p_variables, p_name, p_bindable);
                 GenericMenuPopup.Show(menu, "", Event.current.mousePosition, 240, 300, false, false);
@@ -71,6 +65,8 @@ namespace Dash.Editor
             GUI.color = oldColor;
 
             GUILayout.EndHorizontal();
+
+            return invalidate;
         }
     }
 }