+++ Editor/Scripts/Windows/AOTWindow.cs
@@ -34,7 +34,7 @@ namespace Dash.Editor
         {
             var rect = new Rect(0, 0, position.width, position.height);
             
-            GUICustomUtils.DrawTitle("Dash AOT Scanner/Generator");
+            GUIEditorUtils.DrawTitle("Dash AOT Scanner/Generator");
 
             var titleStyle = new GUIStyle();
             titleStyle.alignment = TextAnchor.MiddleLeft;