+++ Editor/Scripts/Windows/PrefabWindow.cs
@@ -43,7 +43,7 @@ namespace Dash.Editor
         {
             var rect = new Rect(0, 0, position.width, position.height);
             
-            GUICustomUtils.DrawTitle("Dash Prefab Editor");
+            GUIEditorUtils.DrawTitle("Dash Prefab Editor");
 
             var scrollViewStyle = new GUIStyle();
             scrollViewStyle.normal.background = TextureUtils.GetColorTexture(new Color(.1f, .1f, .1f));