+++ CHANGELOG.md
@@ -4,6 +4,38 @@ All notable changes to this project will be documented in this file.
 
 ## RELEASE CANDIDATES
 
+### Release 0.13.4
+
+#### Added
+
+- added support and serialization for prefab modifications on DashVariablesController
+
+#### Changed
+
+- changed DashController initialization now initializes on demand when using Send/Add methods not just on Awake
+- changed DashControllerVariables direct reference type variables now support scene referencing
+- changed removed support for ExposedReference variables [BREAKING]
+- changed ExposedReferences in node properties now cache initial exposedName
+- changed a lot of UI cleanup and changes for variables and node inspector
+- changed custom expression fields now reset their previous value when switching from direct to expression or back
+
+#### Fixed
+
+- fixed setting exposed reference to null not reset its values on all graph usages
+
+### Release 0.13.3 - 27.10.2022
+
+#### Added
+
+- added DashSettingsWindow to have single settings window for various Dash settings
+- added more UI for retarget nodes to show invalid/empty references
+- added speed based tweens support for AnimateAnchoredPositionNode, AnimatePositionNode, AnimateScaleNode, AnimateSizeDeltaNode
+
+#### Changed
+
+- changed minor settings refactors
+
+
 ### Release 0.13.2 - 21.10.2022
 
 #### Added